Changes to buttons

+++ src/edit-a-quiz/EditQuestions.jsx
@@ -7,77 +7,70 @@ const EditQuestions = ({ quiz }) => {
   const nav = useNavigate();
 
   // const questionObject = quiz.questions[index]
-  const [questionObject, setQuestionObject] = useState(quiz.questions[0]);
-  const [question, setQuestion] = useState(questionObject.question);
-  const [confirm, setConfirm] = useState(false);
-  const [correctAnswer, setCorrectAnswer] = useState(
-    questionObject.correctAnswer
-  );
-  const [incAnswersNew, setIncAnswersNew] = useState([]);
-  const [incorrectAnswer1, setIncorrectAnswer1] = useState(
-    questionObject.incorrectAnswers[0]
-  );
-  const [incorrectAnswer2, setIncorrectAnswer2] = useState(
-    questionObject.incorrectAnswers[1]
-  );
-  const [incorrectAnswer3, setIncorrectAnswer3] = useState(
-    questionObject.incorrectAnswers[2]
-  );
-  const [newQuestion, setNewQuestion] = useState(false);
+  
+  const [questionObject, setQuestionObject] = useState(quiz.questions[index])
+  const [question, setQuestion] = useState(questionObject.question)
+  const [confirm, setConfirm] = useState(false)
+  const [correctAnswer, setCorrectAnswer] = useState(questionObject.correctAnswer)
+  const [incAnswersNew, setIncAnswersNew] = useState([])
+  const [incorrectAnswer1, setIncorrectAnswer1] = useState(questionObject.incorrectAnswers[0])
+  const [incorrectAnswer2, setIncorrectAnswer2] = useState(questionObject.incorrectAnswers[1])
+  const [incorrectAnswer3, setIncorrectAnswer3] = useState(questionObject.incorrectAnswers[2])
+  const [newQuestion, setNewQuestion] = useState(false)
+
 
   // this function creates a new question object and put it to the server API
   async function updateQuestion() {
     // creates a new question object to send back to server
     const updatedQuestion = {
-      question,
-      correctAnswer,
-      incorrectAnswers: [incorrectAnswer1, incorrectAnswer2, incorrectAnswer3],
-    };
-
+      question: question, 
+      correctAnswer: correctAnswer,
+      incorrectAnswers: [incorrectAnswer1, incorrectAnswer2, incorrectAnswer3]
+    }
+    console.log(updatedQuestion, 'updated')
     // fetch to API
-    const res = await fetch(
-      `https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`,
-      {
-        method: "PUT",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(updatedQuestion),
-      }
-    );
-
-    const data = await res.json();
-    console.log(data);
-
-    // need to add logic to handle errors
-    alert("Question updated successfully");
+    const res = await fetch(`https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`, {
+      method: 'PUT',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(updatedQuestion)
+    })
   }
 
   // updates index when user click next
-  function handleClickNext(event) {
-    event.preventDefault();
-    setIndex(index + 1);
-    setQuestionObject(quiz.questions[index + 1]);
+  function handleClickSaveNext(event) {
+    event.preventDefault()
+    setIndex(index+1)
+    setQuestionObject(quiz.questions[index+1])
+    setQuestion(quiz.questions[index+1].question)
+    setCorrectAnswer(quiz.questions[index+1].correctAnswer)
+    setIncorrectAnswer1(quiz.questions[index+1].incorrectAnswers[0])
+    setIncorrectAnswer2(quiz.questions[index+1].incorrectAnswers[1])
+    setIncorrectAnswer3(quiz.questions[index+1].incorrectAnswers[2])
+    updateQuestion()
   }
 
   // handle update question
-  function handleUpdate(event) {
-    event.preventDefault();
-    updateQuestion();
+  function handleSubmit(event) {
+    event.preventDefault()
+    updateQuestion()
+    nav('/quizzes')
   }
 
   // updates index when user click next
-  function handleSubmit(event) {
-    event.preventDefault();
-    // updateQuestion()
-    nav("/quizzes");
+  function handleSaveQuestion(event) {
+    event.preventDefault()
+    addQuestion()
+    setNewQuestion(false)
   }
 
   // handle click delete
   function handleClickDelete(event) {
-    event.preventDefault();
-    setConfirm(true);
+    event.preventDefault()
+    setConfirm(true)
+    setIndex(index+1)
   }
 
   // Function to reset the state of the form after submitting a question
@@ -134,21 +127,25 @@ const EditQuestions = ({ quiz }) => {
 
   // handles confirming deletion of a question
   const handleConfirmDelete = async (event) => {
-    event.preventDefault();
-    try {
-      await fetch(
-        `https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`,
-        {
-          method: "DELETE",
-        }
-      );
-      setIndex(index + 1);
-      setConfirm(false);
-    } catch (error) {
-      console.log(error);
+    event.preventDefault()
+    if (quiz.questions.length === 1) {
+      alert('Cannot delete this question. Quiz is required to have at least one question.')
+    } else {
+      try {
+        await fetch(`https://quiz-app-server-production-09e8.up.railway.app/questions/${questionObject._id}`, {
+            method: "DELETE"
+        })
+        setIndex(index+1)
+        setConfirm(false)
+      } catch (error) {
+        console.log(error)
+      }
+      // index === quiz.questions.length-1 ? resetForm() : setIndex(index+1) 
+      setIndex(index+1)
+      setQuestionObject(quiz.questions[index+1])
+      setConfirm(false)``
     }
-  };
-  console.log(questionObject);
+  }
 
   // handles click on delete a question
   const confirmForm = () => {
@@ -156,7 +153,7 @@ const EditQuestions = ({ quiz }) => {
       <>
         <p> Do you want to delete this question? </p>
         <button onClick={handleConfirmDelete}> Confirm </button>
-        <button> Cancel </button>
+        <button onClick={ () => setConfirm(false) }> Cancel </button>
         <br />
       </>
     );
@@ -173,32 +170,21 @@ const EditQuestions = ({ quiz }) => {
       incorrectAnswers: [incorrectAnswer1, incorrectAnswer2, incorrectAnswer3],
     };
     // Post new question to API
-    await fetch(
-      "https://quiz-app-server-production-09e8.up.railway.app/questions",
-      {
-        method: "POST",
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(newQuestion),
-      }
-    );
-    setNewQuestion(false);
-  };
 
-  //render for last question
-  const render = () => {
-    return (
-      <>
-        <br/>
-        <button onClick={handleAddQuestion} > Add a new question </button>
-        <br/>
-        <button onClick={handleSubmit}> Submit </button>
-      </>
-    );
-  };
-  console.log(questionObject._id);
+    const res = await fetch('https://quiz-app-server-production-09e8.up.railway.app/questions', {
+      method: 'POST',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newQuestion)
+    })
+    const data = await res.json()
+    setNewQuestion(false)
+    quiz.questions.push(data)
+    alert('Question added successfully')
+  } 
+
 
   return (
     <>
@@ -230,54 +216,23 @@ const EditQuestions = ({ quiz }) => {
           <div className="incorrect-answers-form d-flex flex-column">
             <label>Incorrect answers:</label>
             <input
-              type="text"
-              defaultValue={questionObject.incorrectAnswers[0]}
-              // value={questionObject.incorrectAnswers[0]}
-              onChange={(e) => setIncorrectAnswer1(e.target.value)}
-            />
-          </div>
-          <input
-            type="text"
-            defaultValue={questionObject.incorrectAnswers[1]}
-            // value={questionObject.incorrectAnswers[1]}
-            onChange={(e) => setIncorrectAnswer2(e.target.value)}
-          />
-          <input
-            type="text"
-            defaultValue={questionObject.incorrectAnswers[2]}
-            // value={questionObject.incorrectAnswers[2]}
-            onChange={(e) => setIncorrectAnswer3(e.target.value)}
-          />
-
-          {confirm && confirmForm()}
-          {newQuestion ? (
-            <button onClick={handleConfirmAdd}> Add question </button>
-          ) : (
-            <>
-            <br/>
-              <div className="d-flex justify-content-between">                
-                <button
-                  onClick={handleUpdate}
-                  
-                >
-                  Save changes
-                </button>
-                {/* <button onClick={ handleAddQuestion }> Add a new question </button> */}
-                <button onClick={handleClickDelete}>
-                  Delete this question
-                </button>
-              </div>
-              {index < quiz.questions.length - 1 ? (
-                <button onClick={handleClickNext}> Next </button>
-              ) : (
-                render()
-              )}
-            </>
-          )}
+              type='text'
+              defaultValue={questionObject.incorrectAnswers[2]}
+              // value={questionObject.incorrectAnswers[2]}
+              onChange={(e) => setIncorrectAnswer3(e.target.value)} 
+              />
+            
         </form>
-      </div>
-    </>
-  );
-};
+        { confirm && confirmForm() }
+        <button onClick={() => nav('/edit-a-quiz')}>Quit</button> 
+        { (index < quiz.questions.length-1 && !confirm) && <button onClick={ handleClickSaveNext }> Save & Next </button> }
+        { newQuestion && <button onClick={ handleConfirmAdd }> Save Question </button> } 
+        { (!confirm && !newQuestion) && <button onClick={ handleClickDelete }> Delete this question </button> }  
+        { (index === quiz.questions.length-1 && !confirm) &&  <button onClick={ handleAddQuestion }> Add a new question </button>}
+        { index === quiz.questions.length-1 && <button onClick={handleSubmit}> Submit </button> }
 
-export default EditQuestions;
+    </>
+  )
+}
+
+export default EditQuestions
